% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kickoffs_functions.R
\name{get_ez_kickoff_league_summary}
\alias{get_ez_kickoff_league_summary}
\title{Get season-to-date summary stats of returns taken out of the end zone at
the league-level}
\usage{
get_ez_kickoff_league_summary(season)
}
\arguments{
\item{season}{An NFL season from 2009-2023}
}
\value{
A tibble kickoff data aggregated to the league-level
}
\description{
Get season-to-date summary stats of returns taken out of the end zone at
the league-level
}
\examples{
# leagie <- get_ez_kickoff_league_summary(2023)
}
