% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kickoffs_functions.R
\name{get_ez_returns_by_team}
\alias{get_ez_returns_by_team}
\title{Get returns taken out of the end zone, by team}
\usage{
get_ez_returns_by_team(season)
}
\arguments{
\item{season}{An NFL season from 2009-2023}
}
\value{
A tibble kickoff data aggregated to the team-level
}
\description{
Get returns taken out of the end zone, by team
}
\examples{
# teams <- get_ez_returns_by_team(2023)
}
