% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kickoffs_functions.R
\name{get_kickoffs}
\alias{get_kickoffs}
\title{Get kickoff play-level data}
\usage{
get_kickoffs(season)
}
\arguments{
\item{season}{An NFL season from 2009-2023}
}
\value{
A tibble of play level kickoff data
}
\description{
Get kickoff play-level data
}
\examples{
# kickoffs <- get_kickoffs(2023)
}
